Refactor: use async await instead of then
