--- src/js/utils/fetchData.js
@@ -1,8 +1,0 @@
-export const fetchData = async url => {
-  try {
-    const data = await fetch(url);
-    return data.json();
-  } catch (err) {
-    console.log(err);
-  }
-};
